Fix 'noImplicitAny' issue with declaration file.
The latest release of grpc-node cannot be used in a typescript project that uses -noImplicitAny.  This is due to a signature being added without a specified return type.  This causes the return type to implicitly be 'any' which TS explicitly disallows.  For projects that use strict flags to catch lots of bugs, this prevents usage of this version of grpc entirely.